Refactor config package
