--- internal/pkg/config/arguments.go
@@ -1,7 +1,0 @@
-package config
-
-type Federation struct {
-	MeshPeers          MeshPeers
-	ExportedServiceSet ExportedServiceSet
-	ImportedServiceSet ImportedServiceSet
-}
