+++ internal/pkg/mcp/import_handler.go
@@ -121,7 +121,7 @@ func (h *ImportedServiceHandler) makeWorkloadEntries(ports []*v1alpha1.ServicePo
 			Ports:   make(map[string]uint32, len(ports)),
 		}
 		for _, p := range ports {
-			we.Ports[p.Name] = h.cfg.MeshPeers.Remote.DataPlane.GetPort()
+			we.Ports[p.Name] = h.cfg.GetRemoteDataPlaneGatewayPort()
 		}
 		workloadEntries = append(workloadEntries, we)
 	}
